refactor truncate functions to separate file
